lib: update Ensembl/RefSeq indexers and builders (include major improvements and sonnar fixes), #TASK-5776, #TASK-5564
