+++ cellbase-lib/src/main/java/org/opencb/cellbase/lib/builders/CellBaseBuilder.java
@@ -18,6 +18,7 @@ package org.opencb.cellbase.lib.builders;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectReader;
+import org.apache.commons.lang3.StringUtils;
 import org.opencb.cellbase.core.exception.CellBaseException;
 import org.opencb.cellbase.core.models.DataSource;
 import org.opencb.cellbase.core.serializer.CellBaseSerializer;
@@ -25,11 +26,14 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
+import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import static org.opencb.cellbase.lib.EtlCommons.*;
 
 /**
  * Created by imedina on 30/08/14.
@@ -39,33 +43,54 @@ public abstract class CellBaseBuilder {
     protected CellBaseSerializer serializer;
     protected ObjectReader dataSourceReader = new ObjectMapper().readerFor(DataSource.class);
 
+    protected boolean checked;
+
     protected Logger logger;
 
+    public static final String CHECKING_BEFORE_BUILDING_LOG_MESSAGE = "Checking files before building {} ...";
+    public static final String CHECKING_DONE_BEFORE_BUILDING_LOG_MESSAGE = "Checking done!";
+
     public static final String BUILDING_LOG_MESSAGE = "Building {} ...";
-    public static final String BUILDING_DONE_LOG_MESSAGE = "Building {} done.";
+    public static final String BUILDING_DONE_LOG_MESSAGE = "Building done!";
 
     public static final String CATEGORY_BUILDING_LOG_MESSAGE = "Building {}/{} ...";
-    public static final String CATEGORY_BUILDING_DONE_LOG_MESSAGE = "Building {}/{} done.";
+    public static final String CATEGORY_BUILDING_DONE_LOG_MESSAGE = "Building done!";
 
     public static final String PARSING_LOG_MESSAGE = "Parsing {} ...";
-    public static final String PARSING_DONE_LOG_MESSAGE = "Parsing {} done.";
+    public static final String PARSING_DONE_LOG_MESSAGE = "Parsing done!";
 
 
     public CellBaseBuilder(CellBaseSerializer serializer) {
         logger = LoggerFactory.getLogger(this.getClass());
 
         this.serializer = serializer;
-        //this.serializer.open();
+        this.checked = false;
     }
 
     public abstract void parse() throws Exception;
 
     public void disconnect() {
-        try {
-            serializer.close();
-        } catch (Exception e) {
-            logger.error("Disconnecting serializer: " + e.getMessage());
+        if (serializer != null) {
+            try {
+                serializer.close();
+            } catch (Exception e) {
+                logger.error("Error closing serializer:\n" + StringUtils.join(e.getStackTrace(), "\n"));
+            }
         }
+    }
+
+    protected List<File> checkFiles(String data, Path downloadPath, int expectedFiles) throws CellBaseException, IOException {
+        return checkFiles(getDataName(data), data, downloadPath, expectedFiles);
+    }
+
+    protected List<File> checkFiles(String label, String data, Path downloadPath, int expectedFiles) throws CellBaseException, IOException {
+        List<File> files = checkFiles(dataSourceReader.readValue(downloadPath.resolve(getDataVersionFilename(data)).toFile()),
+                downloadPath, label);
+        if (files.size() != expectedFiles) {
+            throw new CellBaseException(expectedFiles + " " + label + " files are expected at " + downloadPath + ", but currently there"
+                    + " are " + files.size() + " files");
+        }
+        return files;
     }
 
     protected List<File> checkFiles(DataSource dataSource, Path targetPath, String name) throws CellBaseException {
