+++ cellbase-lib/src/main/java/org/opencb/cellbase/lib/builders/EnsemblGeneBuilderIndexer.java
@@ -16,27 +16,44 @@
 
 package org.opencb.cellbase.lib.builders;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.poi.hssf.usermodel.HSSFSheet;
-import org.apache.poi.hssf.usermodel.HSSFWorkbook;
-import org.apache.poi.ss.usermodel.*;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
+import org.opencb.biodata.formats.feature.mirbase.MirBaseParser;
+import org.opencb.biodata.formats.feature.mirbase.MirBaseParserCallback;
 import org.opencb.biodata.formats.gaf.GafParser;
 import org.opencb.biodata.formats.io.FileFormatException;
+import org.opencb.biodata.models.core.FeatureOntologyTermAnnotation;
+import org.opencb.biodata.models.core.MiRnaGene;
+import org.opencb.biodata.models.core.MirnaTarget;
 import org.opencb.biodata.models.core.Xref;
-import org.opencb.biodata.models.core.*;
-import org.opencb.biodata.models.variant.avro.*;
+import org.opencb.biodata.models.variant.avro.Constraint;
+import org.opencb.biodata.models.variant.avro.Expression;
+import org.opencb.biodata.models.variant.avro.ExpressionCall;
+import org.opencb.cellbase.core.exception.CellBaseException;
 import org.opencb.commons.utils.FileUtils;
+import org.rocksdb.RocksDB;
 import org.rocksdb.RocksDBException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-import java.io.*;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.zip.GZIPInputStream;
 
-public class EnsemblGeneBuilderIndexer extends GeneBuilderIndexer{
+import static org.opencb.cellbase.lib.EtlCommons.ENSEMBL_DATA;
+import static org.opencb.cellbase.lib.builders.CellBaseBuilder.PARSING_DONE_LOG_MESSAGE;
+import static org.opencb.cellbase.lib.builders.CellBaseBuilder.PARSING_LOG_MESSAGE;
+
+public class EnsemblGeneBuilderIndexer extends GeneBuilderIndexer {
 
     private static final String DESCRIPTION_SUFFIX = "_description";
     private static final String XREF_SUFFIX = "_xref";
@@ -56,12 +73,12 @@ public class EnsemblGeneBuilderIndexer extends GeneBuilderIndexer{
                       Path proteinFastaFile, Path cDnaFastaFile, String species, Path geneExpressionFile, Path geneDrugFile, Path hpoFile,
                       Path disgenetFile, Path gnomadFile, Path geneOntologyAnnotationFile, Path miRBaseFile, Path miRTarBaseFile,
                       Path cancerGeneGensusFile, Path cancerHostpotFile, Path canonicalFile, Path tso500File, Path eglhHaemOncFile)
-            throws IOException, RocksDBException, FileFormatException {
-        indexDescriptions(geneDescriptionFile);
-        indexXrefs(xrefsFile, uniprotIdMappingFile);
+            throws IOException, RocksDBException, FileFormatException, CellBaseException {
+//        indexDescriptions(geneDescriptionFile);
+//        indexXrefs(xrefsFile, uniprotIdMappingFile);
         indexHgncIdMapping(hgncFile);
-        indexManeMapping(maneFile, "ensembl");
-        indexLrgMapping(lrgFile, "ensembl");
+        indexManeMapping(maneFile, ENSEMBL_DATA);
+        indexLrgMapping(lrgFile, ENSEMBL_DATA);
         indexProteinSequences(proteinFastaFile);
         indexCdnaSequences(cDnaFastaFile);
         indexExpression(species, geneExpressionFile);
@@ -69,13 +86,13 @@ public class EnsemblGeneBuilderIndexer extends GeneBuilderIndexer{
         indexDiseases(hpoFile, disgenetFile);
         indexConstraints(gnomadFile);
         indexOntologyAnnotations(geneOntologyAnnotationFile);
-        indexMiRBase(miRBaseFile);
+        indexMiRBase(species, miRBaseFile);
         indexMiRTarBase(miRTarBaseFile);
-        indexCancerGeneCensus(cancerGeneGensusFile);
+//        indexCancerGeneCensus(cancerGeneGensusFile);
         indexCancerHotspot(cancerHostpotFile);
-        indexCanonical(canonicalFile);
-        indexTSO500(tso500File);
-        indexEGLHHaemOnc(eglhHaemOncFile);
+//        indexCanonical(canonicalFile);
+//        indexTSO500(tso500File);
+//        indexEGLHHaemOnc(eglhHaemOncFile);
     }
 
     private void indexDescriptions(Path geneDescriptionFile) throws IOException, RocksDBException {
@@ -233,129 +250,6 @@ public class EnsemblGeneBuilderIndexer extends GeneBuilderIndexer{
         return rocksDbManager.getExpression(rocksdb, key);
     }
 
-    private void indexDrugs(Path geneDrugFile) throws IOException, RocksDBException {
-        if (geneDrugFile != null && Files.exists(geneDrugFile) && Files.size(geneDrugFile) > 0) {
-            logger.info("Loading gene-drug interaction data from '{}'", geneDrugFile);
-            BufferedReader br = FileUtils.newBufferedReader(geneDrugFile);
-
-            // Skip header
-            br.readLine();
-
-            int lineCounter = 1;
-            String line;
-            String currentGene = "";
-            List<GeneDrugInteraction> drugs = new ArrayList<>();
-            while ((line = br.readLine()) != null) {
-                String[] parts = line.split("\t");
-                String geneName = parts[0];
-                if (currentGene.equals("")) {
-                    currentGene = geneName;
-                } else if (!currentGene.equals(geneName)) {
-                    rocksDbManager.update(rocksdb, currentGene + DRUGS_SUFFIX, drugs);
-                    drugs = new ArrayList<>();
-                    currentGene = geneName;
-                }
-
-                String source = null;
-                if (parts.length >= 4) {
-                    source = parts[3];
-                }
-
-                String interactionType = null;
-                if (parts.length >= 5) {
-                    interactionType = parts[4];
-                }
-
-                String drugName = null;
-                if (parts.length >= 8) {
-                    // if drug name column is empty, use drug claim name instead
-                    drugName = StringUtils.isEmpty(parts[7]) ? parts[6] : parts[7];
-                }
-                if (StringUtils.isEmpty(drugName)) {
-                    // no drug name
-                    continue;
-                }
-
-                String chemblId = null;
-                if (parts.length >= 9) {
-                    chemblId = parts[8];
-                }
-
-                List<String> publications = new ArrayList<>();
-                if (parts.length >= 10 && parts[9] != null) {
-                    publications = Arrays.asList(parts[9].split(","));
-                }
-
-                GeneDrugInteraction drug = new GeneDrugInteraction(
-                        geneName, drugName, source, null, null, interactionType, chemblId, publications);
-                drugs.add(drug);
-                lineCounter++;
-            }
-            br.close();
-            // update last gene
-            rocksDbManager.update(rocksdb, currentGene + DRUGS_SUFFIX, drugs);
-        } else {
-            logger.warn("Gene drug file " + geneDrugFile + " not found");
-            logger.warn("Ignoring " + geneDrugFile);
-        }
-    }
-
-    public List<GeneDrugInteraction> getDrugs(String id) throws RocksDBException, IOException {
-        String key = id + DRUGS_SUFFIX;
-        return rocksDbManager.getDrugs(rocksdb, key);
-    }
-
-    private void indexDiseases(Path hpoFilePath, Path disgenetFilePath) throws IOException, RocksDBException {
-        Map<String, List<GeneTraitAssociation>> geneDiseaseAssociationMap = new HashMap<>(50000);
-        String line;
-
-        if (hpoFilePath != null && hpoFilePath.toFile().exists() && Files.size(hpoFilePath) > 0) {
-            try (BufferedReader bufferedReader = FileUtils.newBufferedReader(hpoFilePath)) {
-                // skip first header line
-                bufferedReader.readLine();
-                while ((line = bufferedReader.readLine()) != null) {
-                    String[] fields = line.split("\t");
-                    String omimId = fields[6];
-                    String geneSymbol = fields[3];
-                    String hpoId = fields[0];
-                    String diseaseName = fields[1];
-                    GeneTraitAssociation disease =
-                            new GeneTraitAssociation(omimId, diseaseName, hpoId, 0f, 0, new ArrayList<>(), new ArrayList<>(), "hpo");
-                    addValueToMapElement(geneDiseaseAssociationMap, geneSymbol, disease);
-                }
-            }
-        }
-
-        if (disgenetFilePath != null && disgenetFilePath.toFile().exists() && Files.size(disgenetFilePath) > 0) {
-            try (BufferedReader bufferedReader = FileUtils.newBufferedReader(disgenetFilePath)) {
-                // skip first header line
-                bufferedReader.readLine();
-                while ((line = bufferedReader.readLine()) != null) {
-                    String[] fields = line.split("\t");
-                    String diseaseId = fields[4];
-                    String diseaseName = fields[5];
-                    String score = fields[9];
-                    String numberOfPubmeds = fields[13].trim();
-                    String numberOfSNPs = fields[14];
-                    String source = fields[15];
-                    GeneTraitAssociation disease = new GeneTraitAssociation(diseaseId, diseaseName, "", Float.parseFloat(score),
-                            Integer.parseInt(numberOfPubmeds), Collections.singletonList(numberOfSNPs), Collections.singletonList(source),
-                            "disgenet");
-                    addValueToMapElement(geneDiseaseAssociationMap, fields[1], disease);
-                }
-            }
-        }
-
-        for (Map.Entry<String, List<GeneTraitAssociation>> entry : geneDiseaseAssociationMap.entrySet()) {
-            rocksDbManager.update(rocksdb, entry.getKey() + DISEASE_SUFFIX, entry.getValue());
-        }
-    }
-
-    public List<GeneTraitAssociation> getDiseases(String id) throws RocksDBException, IOException {
-        String key = id + DISEASE_SUFFIX;
-        return rocksDbManager.getDiseases(rocksdb, key);
-    }
-
     private void indexConstraints(Path gnomadFile) throws IOException, RocksDBException {
         if (gnomadFile != null && Files.exists(gnomadFile) && Files.size(gnomadFile) > 0) {
             logger.info("Loading OE scores from '{}'", gnomadFile);
@@ -384,7 +278,7 @@ public class EnsemblGeneBuilderIndexer extends GeneBuilderIndexer{
                 rocksDbManager.update(rocksdb, transcriptIdentifier + CONSTRAINT_SUFFIX, constraints);
 
                 if ("TRUE".equalsIgnoreCase(canonical)) {
-                     rocksDbManager.update(rocksdb, geneIdentifier + CONSTRAINT_SUFFIX, constraints);
+                    rocksDbManager.update(rocksdb, geneIdentifier + CONSTRAINT_SUFFIX, constraints);
                 }
             }
             br.close();
@@ -432,66 +326,13 @@ public class EnsemblGeneBuilderIndexer extends GeneBuilderIndexer{
         return rocksDbManager.getOntologyAnnotations(rocksdb, key);
     }
 
-    private void indexMiRBase(Path miRBaseFile) throws IOException, RocksDBException {
-        if (miRBaseFile != null && Files.exists(miRBaseFile) && Files.size(miRBaseFile) > 0) {
-            logger.info("Loading mirna from '{}'", miRBaseFile);
-            FileInputStream fileInputStream = new FileInputStream(miRBaseFile.toFile());
-            HSSFWorkbook workbook = new HSSFWorkbook(fileInputStream);
-            HSSFSheet sheet = workbook.getSheetAt(0);
-            Iterator<org.apache.poi.ss.usermodel.Row> iterator = sheet.iterator();
-            while (iterator.hasNext()) {
-                Row currentRow = iterator.next();
-                Iterator<Cell> cellIterator = currentRow.iterator();
+    private void indexMiRBase(String species, Path miRBaseFile) throws IOException {
+        logger.info(PARSING_LOG_MESSAGE, miRBaseFile);
 
-                org.apache.poi.ss.usermodel.Cell cell = cellIterator.next();
-                String miRBaseAccession = cell.getStringCellValue();
+        MirBaseCallback callback = new MirBaseCallback(rocksdb, rocksDbManager);
+        MirBaseParser.parse(miRBaseFile, species, callback);
 
-                cell = cellIterator.next();
-                String miRBaseID = cell.getStringCellValue();
-
-                cell = cellIterator.next();
-                String status = cell.getStringCellValue();
-
-                cell = cellIterator.next();
-                String sequence = cell.getStringCellValue();
-
-                cell = cellIterator.next();
-                String mature1Accession = cell.getStringCellValue();
-
-                cell = cellIterator.next();
-                String mature1Id = cell.getStringCellValue();
-
-                cell = cellIterator.next();
-                String mature1Sequence = cell.getStringCellValue();
-
-                String mature2Accession = "";
-                String mature2Id = "";
-                String mature2Sequence = "";
-                if (cellIterator.hasNext()) {
-                    cell = cellIterator.next();
-                    mature2Accession = cell.getStringCellValue();
-
-                    cell = cellIterator.next();
-                    mature2Id = cell.getStringCellValue();
-
-                    cell = cellIterator.next();
-                    mature2Sequence = cell.getStringCellValue();
-                }
-
-                MiRnaGene miRNAGene = new MiRnaGene(miRBaseAccession, miRBaseID, status, sequence, new ArrayList<>());
-                int cdnaStart = sequence.indexOf(mature1Sequence);
-                int cdnaEnd = cdnaStart + mature1Sequence.length();
-                miRNAGene.addMiRNAMature(mature1Accession, mature1Id, mature1Sequence, cdnaStart, cdnaEnd);
-
-                cdnaStart = sequence.indexOf(mature2Sequence);
-                cdnaEnd = cdnaStart + mature2Sequence.length();
-                miRNAGene.addMiRNAMature(mature2Accession, mature2Id, mature2Sequence, cdnaStart, cdnaEnd);
-
-                rocksDbManager.update(rocksdb, miRBaseID + MIRBASE_SUFFIX, miRNAGene);
-            }
-        } else {
-            logger.error("mirna file not found");
-        }
+        logger.info(PARSING_DONE_LOG_MESSAGE, miRBaseFile);
     }
 
     public MiRnaGene getMirnaGene(String transcriptId) throws RocksDBException, IOException {
@@ -509,115 +350,9 @@ public class EnsemblGeneBuilderIndexer extends GeneBuilderIndexer{
         return null;
     }
 
-    private void indexMiRTarBase(Path miRTarBaseFile) throws IOException, RocksDBException {
-        if (miRTarBaseFile != null && Files.exists(miRTarBaseFile) && Files.size(miRTarBaseFile) > 0) {
-            logger.info("Loading mirna targets from '{}'", miRTarBaseFile);
-            FileInputStream file = new FileInputStream(miRTarBaseFile.toFile());
-            Workbook workbook = new XSSFWorkbook(file);
-            Sheet sheet = workbook.getSheetAt(0);
-            Iterator<Row> iterator = sheet.iterator();
-            String currentMiRTarBaseId = null;
-            String currentMiRNA = null;
-            String currentGene = null;
-            List<TargetGene> targetGenes = new ArrayList<>();
-            Map<String, List<MirnaTarget>> geneToMirna = new HashMap<>();
-            while (iterator.hasNext()) {
-                Row currentRow = iterator.next();
-
-                Iterator<Cell> cellIterator = currentRow.iterator();
-                Cell cell = cellIterator.next();
-
-                // Iterate columns
-                String miRTarBaseId = cell.getStringCellValue();
-
-                // skip header
-                if (miRTarBaseId.startsWith("miRTarBase")) {
-                    continue;
-                }
-
-                if (currentMiRTarBaseId == null) {
-                    currentMiRTarBaseId = miRTarBaseId;
-                }
-
-                cell = cellIterator.next();
-                String miRNA = cell.getStringCellValue();
-                if (currentMiRNA == null) {
-                    currentMiRNA = miRNA;
-                }
-
-                // Skip species
-                cellIterator.next();
-
-                // Read target gene
-                cell = cellIterator.next();
-                String geneName = cell.getStringCellValue();
-                if (currentGene == null) {
-                    currentGene = geneName;
-                }
-
-                // Skip entrez gene
-                cellIterator.next();
-                // Skip species
-                cellIterator.next();
-
-                if (!miRTarBaseId.equals(currentMiRTarBaseId) || !geneName.equals(currentGene)) {
-                    // new entry, store current one
-                    MirnaTarget miRnaTarget = new MirnaTarget(currentMiRTarBaseId, "miRTarBase", currentMiRNA, targetGenes);
-                    addValueToMapElement(geneToMirna, currentGene, miRnaTarget);
-                    targetGenes = new ArrayList<>();
-                    currentGene = geneName;
-                    currentMiRTarBaseId = miRTarBaseId;
-                    currentMiRNA = miRNA;
-                }
-
-                // experiment
-                cell = cellIterator.next();
-                String experiment = cell.getStringCellValue();
-
-                // support type
-                cell = cellIterator.next();
-                String supportType = cell.getStringCellValue();
-
-                // pubmed
-                cell = cellIterator.next();
-                String pubmed;
-                // seems to vary, so check both
-                if (cell.getCellType().equals(CellType.NUMERIC)) {
-//                    pubmed = String.valueOf(cell.getNumericCellValue());
-                    pubmed = Integer.toString(Double.valueOf(cell.getNumericCellValue()).intValue());
-                } else {
-                    pubmed = cell.getStringCellValue();
-                }
-
-                targetGenes.add(new TargetGene(experiment, supportType, pubmed));
-            }
-
-            // parse last entry
-            MirnaTarget miRnaTarget = new MirnaTarget(currentMiRTarBaseId, "miRTarBase", currentMiRNA,
-                    targetGenes);
-            addValueToMapElement(geneToMirna, currentGene, miRnaTarget);
-
-            for (Map.Entry<String, List<MirnaTarget>> entry : geneToMirna.entrySet()) {
-                rocksDbManager.update(rocksdb, entry.getKey() + MIRTARBASE_SUFFIX, entry.getValue());
-            }
-        } else {
-            logger.error("mirtarbase file not found");
-        }
-    }
-
     public List<MirnaTarget> getMirnaTargets(String geneName) throws RocksDBException, IOException {
         String key = geneName + MIRTARBASE_SUFFIX;
         return rocksDbManager.getMirnaTargets(rocksdb, key);
-    }
-
-    private static <T> void addValueToMapElement(Map<String, List<T>> map, String key, T value) {
-        if (map.containsKey(key)) {
-            map.get(key).add(value);
-        } else {
-            List<T> valueList = new ArrayList<>();
-            valueList.add(value);
-            map.put(key, valueList);
-        }
     }
 
     protected void indexCanonical(Path canonocalFile) throws IOException, RocksDBException {
@@ -651,5 +386,31 @@ public class EnsemblGeneBuilderIndexer extends GeneBuilderIndexer{
             return null;
         }
         return new String(bytes);
+    }
+
+    // Implementation of the MirBaseParserCallback function
+    public class MirBaseCallback implements MirBaseParserCallback {
+
+        private RocksDB rocksDB;
+        private RocksDbManager rocksDbManager;
+        private Logger logger;
+
+        public MirBaseCallback(RocksDB rocksDB, RocksDbManager rocksDbManager) {
+            this.rocksDB = rocksDB;
+            this.rocksDbManager = rocksDbManager;
+            this.logger = LoggerFactory.getLogger(this.getClass());
+        }
+
+        @Override
+        public boolean processMiRnaGene(MiRnaGene miRnaGene) {
+            try {
+                rocksDbManager.update(rocksdb, miRnaGene.getId() + MIRBASE_SUFFIX, miRnaGene);
+            } catch (JsonProcessingException | RocksDBException e) {
+                logger.warn("Something wrong happened when processing miRNA gene {}: {}", miRnaGene.getId(),
+                        StringUtils.join(e.getStackTrace(), "\t"));
+                return false;
+            }
+            return true;
+        }
     }
 }
